requirements.txt
pandas==2.2.0
requests==2.31.0
plotly==5.18.0
python-dotenv==1.0.1
openpyxl==3.1.2
urllib3==2.2.0
